Cleaned test code
