--- fund_test.go
@@ -1,33 +1,0 @@
-package alipay
-
-//func TestAliPay_FundTransToAccountTransfer(t *testing.T) {
-//	fmt.Println("========== AliPayFundTransToAccountTransfer ==========")
-//	var p = AliPayFundTransToAccountTransfer{}
-//	p.OutBizNo = "xxxx"
-//	p.PayeeType = "ALIPAY_LOGONID"
-//	p.PayeeAccount = "xwmkjn7612@sandbox.com"
-//	p.Amount = "100"
-//	fmt.Println(client.FundTransToAccountTransfer(p))
-//}
-
-//func TestAliPay_FundAuthOrderVoucherCreate(t *testing.T) {
-//	fmt.Println("========== AliPayFundAuthOrderVoucherCreate ==========")
-//	var p = AliPayFundAuthOrderVoucherCreate{}
-//	p.OutOrderNo = "1111"
-//	p.OutRequestNo = "222"
-//	p.OrderTitle = "eee"
-//	p.Amount = "1001"
-//	fmt.Println(client.FundAuthOrderVoucherCreate(p))
-//}
-//
-//func TestAliPay_FundAuthOrderAppFreeze(t *testing.T) {
-//	fmt.Println("========== AliPayFundAuthOrderAppFreeze ==========")
-//	var p = AliPayFundAuthOrderAppFreeze{}
-//	p.OutOrderNo = "111"
-//	p.OutRequestNo = "xxxxx"
-//	p.OrderTitle = "test"
-//	p.Amount = "100"
-//	p.ProductCode = "PRE_AUTH_ONLINE"
-//
-//	fmt.Println(client.FundAuthOrderAppFreeze(p))
-//}
